Virtual Environment
Making a virutal environment to run this code in.

The file requirements.txt has all of the dependencies in it and can be used with pip to make a duplicate environment.

Put the .env/ folder in the .gitignore file so that it will not be saved to the repo.

Removed the Add_earth command becasue I remembered that Boulder didn't want it.